pyproject.toml
[project]
name = "betterai-project"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
  "networkx>=3.4.2",
  "numpy>=1.26.4",
  "tqdm>=4.67.1",
  "requests>=2.32.5",
  "emoji>=2.15.0",
  "google-api-python-client>=2.185.0",
  "torch>=2.2.2",
  "rdflib>=7.3.0",
  "attrs>=25.4.0",
  "ipykernel>=7.0.1",
  # "rdflib-hdt>=3.1",  # Temporarily disabled - doesn't compile on Windows
  "spacy>=3.8.7",
  "spacy-transformers>=1.3.9",
  "transformers>=4.49.0",
  "ai21>=3.0.0",
  "psycopg2-binary>=2.9.9",
  "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "isort>=7.0.0",
    "pytest>=8.4.2",
    "ruff>=0.14.2",
]

[tool.uv]
required-environments=["sys_platform == 'darwin' and platform_machine == 'x86_64'"]



############################
# Code linting
############################
[tool.black]
line-length = 100
target-version = ['py313']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
skip = []

[tool.ruff]
target-version = "py313"
exclude = ['__pycache__',]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "UP045", # Use `X | None` for type annotations
    "UP031", # Allow % symbol for formatting, aligns with django conventions
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true





